Update protocol V1 command and packet handling for incubation tables
- Update `RequestTableCommandV1` to use `number` instead of `deviceNumber`
- Implement `SetTableCommandV1` and `TablePacketV1` to handle a list of `TableDayV1` objects representing 30 days of incubation data
- Implement `TableDayV1Serializer` and `TableDayV1Parser` for handling 8-byte daily data structures (including temperature, humidity, and ventilation parameters)
- Fix `ConfirmPacketV1` to parse `commandId` as a 1-byte value (u8) instead of 2-byte (u16)
- Update various V1 serializers and parsers to clarify that `commandId` is treated as a 1-byte field in the protocol documentation/comments
- Complete `TablePacketV1Parser` implementation to parse the full 240-byte payload into a list of daily settings

+++ data/repos/src/commonMain/kotlin/ua/graviton/isida/data/parsers/v1/ConfirmPacketV1Parser.kt
@@ -7,12 +7,12 @@ import ua.graviton.isida.data.protocol.packets.v1.ConfirmPacketV1
 object ConfirmPacketV1Parser : PacketParser<ConfirmPacketV1> {
     override val version: Int = 1       // [0x01]
     override val length: Int = 2        // [0x02, 0x00]
-    override val commandId: Int = 91    // [0x5B, 0x00]
+    override val commandId: Int = 91    // [0x5B]
 
     override fun parsePayload(reader: PacketReader): Result<ConfirmPacketV1> = Result.runCatching {
         require(reader.data.size == length) { "Data size must be $length but was ${reader.data.size}" }
         ConfirmPacketV1(
-            commandId = reader.u16(),
+            commandId = reader.u8(),
         )
     }
 }
